user delete on manifest,receipt

+++ conti/src/main/java/com/conti/setting/usercontrol/UserRestController.java
@@ -50,6 +50,8 @@ import org.springframework.web.util.UriComponentsBuilder;
 import com.conti.config.SessionListener;
 import com.conti.hsn.Hsn;
 import com.conti.hsn.HsnDao;
+import com.conti.manifest.ManifestDao;
+import com.conti.manifest.ManifestModel;
 import com.conti.master.branch.BranchDao;
 import com.conti.master.branch.BranchModel;
 import com.conti.master.customer.CustomerDao;
@@ -69,6 +71,8 @@ import com.conti.others.DateTimeCalculation;
 import com.conti.others.Loggerconf;
 import com.conti.others.SendMailSMS;
 import com.conti.others.UserInformation;
+import com.conti.receipt.ReceiptDao;
+import com.conti.receipt.ReceiptModel;
 import com.conti.settings.company.Company;
 import com.conti.settings.company.CompanySettingDAO;
 import com.conti.settings.price.PriceSetting;
@@ -124,6 +128,10 @@ public class UserRestController {
 	private VehicleDao vehicleDao;
 	@Autowired 
 	private HsnDao hsnDao;
+	@Autowired
+	private ManifestDao mDao;
+	@Autowired
+	private ReceiptDao receiptDao;
 	
 	Loggerconf loggerconf = new Loggerconf();
 	ConstantValues constantVal = new ConstantValues();
@@ -412,6 +420,8 @@ public class UserRestController {
 		ServiceMaster serviceMaster = serviceDao.getUserid(id, id);
 		VehicleMaster vehicleMaster = vehicleDao.getUserId(id, id);
 		Hsn hsn = hsnDao.getUserId(id, id);
+		ManifestModel manifest = mDao.getUserId(id, id);
+		ReceiptModel receipt = receiptDao.getUser(id, id);
 		
 		try {
 			User currentUser = usersDao.get(id);
@@ -419,7 +429,8 @@ public class UserRestController {
 				return new ResponseEntity<String>(HttpStatus.NOT_FOUND);
 			} else {
 				if(company == null && priceSetting == null && branchModel == null && customerModel == null && employeeMaster == null
-						&& location == null && product == null && serviceMaster == null && vehicleMaster == null && hsn == null) {
+						&& location == null && product == null && serviceMaster == null && vehicleMaster == null && hsn == null
+						&& manifest == null && receipt == null) {
 					usersDao.delete(id);
 					loggerconf.saveLogger(username, request.getServletPath(), ConstantValues.DELETE_SUCCESS, null);
 					return new ResponseEntity<String>( HttpStatus.OK);
